% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/chromunity.R
\name{annotate}
\alias{annotate}
\title{Annoatate bins with counts and covariates}
\usage{
annotate(
  binsets,
  concatemers,
  covariates = NULL,
  k = 5,
  interchromosomal.dist = 1e+08,
  interchromosomal.table = NULL,
  gg = NULL,
  mc.cores = 5,
  numchunks = 200 * mc.cores - 1,
  seed = 42,
  verbose = TRUE,
  unique.per.setid = TRUE,
  resolution = 50000
)
}
\arguments{
\item{binsets}{GRanges of bins with fields $bid specifying binset id}

\item{concatemers}{GRanges of monomers with fields seqnames, start, end, and $cid specifying concatemer id, which will be counted across each binset}

\item{covariates}{Covariate object specifying numeric or interval covariates to merge with this binset}

\item{k}{the max cardinality of sets to annotate from the power set of each binmer}

\item{interchromosomal.dist}{numeric scalar of "effective" distance for inter chromosomal bins [1e8]}

\item{gg}{optional gGraph input specifying alternate distance function}

\item{mc.cores}{integer how many cores to parallelize}

\item{verbose}{logical flag}
}
\value{
data.table of sub-binsets i.e. k-power set of binsets annotated with $count field representing covariates, ready for fitting, **one row per binset
}
\description{
Given n binsets annotates their k-power set (i.e. all sets in the power set with cardinality <= k) with
basic (min median max distance, cardinality, width) and optional numerical and interval covariates provided as input.
}
\author{
Aditya Deshpande, Marcin Imielinski
}
