% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/chromunity.R
\name{annotate_multimodal_clusters}
\alias{annotate_multimodal_clusters}
\title{Annotates communities that are very dense with respect to genomic coordinates.}
\usage{
annotate_multimodal_clusters(granges, which.gr, min.memb = 50)
}
\arguments{
\item{granges}{GRanges output from chromunity function}

\item{which.gr}{the GRanges for the window of interest}

\item{min.memb}{numeric minimum number of members needed to be in a community to be considered for further analyses}

\item{k.min}{numeric the threshold to number of concatemers pairs to be considered "similar"}

\item{tiles}{GRanges object dividing the genome in a fixed size bin to be defined by user}
}
\value{
\code{annotate_local_clusters} returns input GRanges with additional columns as follows:

   \item{multimodal}{  boolean; \cr
             whether a community had multimodal contacts based on parameters set by user
   }
}
\description{
Using the nature of distribution of contacts on genomic coordinates, annotates community that are local
}
\author{
Aditya Deshpande
}
