% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/chromunity.R
\name{background}
\alias{background}
\title{Generate random bins}
\usage{
background(
  binsets,
  n = nrow(binsets),
  pseudocount = 1,
  resolution = 50000,
  gg = NULL,
  interchromosomal.dist = 1e+08,
  interchromosomal.table = NULL,
  verbose = TRUE,
  mc.cores = 5
)
}
\arguments{
\item{binsets}{GRanges of bins with fields seqnames, start, end, and $bid specifying binset id}

\item{n}{integer scalar specifying number of sets to sample (default = nrow(binsets))}

\item{pseudocount}{integer used to populate chr-chr transition matrix}

\item{resolution}{integer resolution to use for simulating bin-sets}

\item{gg}{gGnome object to be used for simulating bin-sets in the presence of structural variation}

\item{interchromosomal.dist}{integer inferred average inter-chr distance to be used in case of whole genome simulation}

\item{interchromosomal.table}{data.table optional table that contains specific chr-chr inferred distances.}

\item{verbose}{boolean "verbose" flag (default = FALSE)}

\item{mc.cores}{integer Number of cores in mclapply (default = 10)}
}
\description{
Given n binsets generates random "background" binsets that mirrors the input binset characteristics with respect to chromosome, width, and distance.

Generates (chromosome specific) kernels for width, cardinality, distance and models interchromosomal binsets by allowing
a chromosome transition matrix which then lands you on a coordinate that is also chosen from a kernel. 
x
}
\author{
Aditya Deshpande, Marcin Imielinski
}
