% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/chromunity.R
\name{c.Cov}
\alias{c.Cov}
\title{title}
\usage{
\method{c}{Cov}(...)
}
\arguments{
\item{...}{A series of Covariates, note all objects must be of type Cov}
}
\value{
Cov object that can be passed directly into the FishHook object constructor that contains all of the Cov covariates
Passed in the ... param
}
\description{
Override the c operator for covariates so that you can merge them like a vector
}
\author{
Zoran Z. Gajic
}
