% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/chromunity.R
\name{ChromObj}
\alias{ChromObj}
\title{Chrom object}
\description{
Vectorized object that stores the output of chromunity analysis and can be inputted into annotate / synergy functions.
The main accessors are $binsets and $concatemers which each return GRanges linked through a chromunity id $chid field

Chromunities can be subsetted, concatenated.  Concatenation will result in deduping any $chid that are present in two inputs.
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-new}{\code{ChromObj$new()}}
\item \href{#method-subset}{\code{ChromObj$subset()}}
\item \href{#method-print}{\code{ChromObj$print()}}
\item \href{#method-gtrack}{\code{ChromObj$gtrack()}}
\item \href{#method-gm}{\code{ChromObj$gm()}}
\item \href{#method-clone}{\code{ChromObj$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-new"></a>}}
\if{latex}{\out{\hypertarget{method-new}{}}}
\subsection{Method \code{new()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{ChromObj$new(
  binsets = GRanges(),
  concatemers = GRanges(),
  meta = data.table(),
  verbose = TRUE
)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-subset"></a>}}
\if{latex}{\out{\hypertarget{method-subset}{}}}
\subsection{Method \code{subset()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{ChromObj$subset(ix, ...)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-print"></a>}}
\if{latex}{\out{\hypertarget{method-print}{}}}
\subsection{Method \code{print()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{ChromObj$print(...)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-gtrack"></a>}}
\if{latex}{\out{\hypertarget{method-gtrack}{}}}
\subsection{Method \code{gtrack()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{ChromObj$gtrack(
  name = "",
  binsets = TRUE,
  concatemers = TRUE,
  heatmap = FALSE,
  binsize = 10000,
  clim = NA,
  ...
)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-gm"></a>}}
\if{latex}{\out{\hypertarget{method-gm}{}}}
\subsection{Method \code{gm()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{ChromObj$gm(binsize = 5000)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-clone"></a>}}
\if{latex}{\out{\hypertarget{method-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{ChromObj$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
