% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/chromunity.R
\name{chromunity}
\alias{chromunity}
\title{Detect communitites and make bin-sets}
\usage{
chromunity(
  concatemers,
  resolution = 50000,
  region = si2gr(concatemers),
  windows = NULL,
  piecewise = TRUE,
  shave = FALSE,
  bthresh = 3,
  cthresh = 3,
  max.size = 2^31 - 1,
  subsample.frac = NULL,
  window.size = 2e+06,
  max.slice = 1e+06,
  min.support = 5,
  stride = window.size/2,
  mc.cores = 5,
  k.knn = 25,
  k.min = 5,
  pad = 1000,
  peak.thresh = 0.85,
  seed = 42,
  verbose = TRUE
)
}
\arguments{
\item{concatemers}{GRanges with $cid}

\item{resolution}{bin size for community detection [5e4]}

\item{region}{region to run on [si2gr(concatemers)]}

\item{windows}{GRanges or GRangesList of windows to test, more appropriate for testing specific windows, e.g. targets such as promoters/enhancers windows. If sliding window approach is desired, keep this parameter as NULL and use piecewise argument.}

\item{piecewise}{logical flag specifying whether to run one window at a time ie to cluster concatemers that (partially overlap) that window, note: this may}

\item{shave}{logical flag specifying whether to iteratively "shave" concatemers and bins to a subset C and B where every bin in B has at leat bthresh concatemer support across C and every concatemer in C has order / cardinality of at least cthresh across B, this is done by iteratively removing bins and concatemers until you reach a fixed point}

\item{max.size}{max size of problem (concatemers x bins) to consider (default is 2^31-1).  If we hit this problem size, will subsample concatemers so that concatemers * bins is < max.dim, this step is done downstream of shaving}

\item{window.size}{window size to do community detection within}

\item{k.min}{minimal number of nearest neighbors an edge in KNN graph needs to have before community detection}

\item{pad}{integer pad to use when computing the footprint of each chromunity and finding peak regions which become binsets}

\item{peak.thresh}{peak threshold with which to call a peak}

\item{tiles.k.knn}{KNN parameter specifying how many nearest neighbors to sample when building KNN graph}
}
\value{
list with items $binset,  $support, $params: $binsets is GRanges of bins with field $bid corresponding to binset id and $support which is the concatemer community supporting the binset which are GRanges with $bid
}
\description{
Runs genome-wide chromunity detection across a sliding or provided genomic window
}
\author{
Aditya Deshpande, Marcin Imielinski
}
