% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/chromunity.R
\name{concatemer_communities}
\alias{concatemer_communities}
\title{Chromunity sub-function}
\usage{
concatemer_communities(
  concatemers,
  k.knn = 25,
  k.min = 5,
  drop.small = FALSE,
  small = 10000,
  max.size = 2^31 - 1,
  subsample.frac = NULL,
  seed = 42,
  verbose = TRUE,
  debug = FALSE
)
}
\arguments{
\item{concatemers}{GRanges of monomers with field $cid indicating concatemer id and $binid represent bin id}

\item{k.min}{minimal number of nearest neighbors an edge in KNN graph needs to have before community detection}

\item{drop.small}{logical flag specifying whether to remove "small" concatemers ie those with a footprint <= small argument [FALSE]}

\item{small}{integer threshold for bases that define small concatemers, only relevant if drop.small = TRUE}

\item{subsample.frac}{optional arg specifying fraction of concatemers to subsample [NULL]}

\item{seed}{seed for subsampling}

\item{tiles.k.knn}{KNN parameter specifying how many nearest neighbors to sample when building KNN graph}
}
\value{
GRanges of concatemers labeled by $c mmunity which specifies community id
}
\description{
Low level function that labels concatemers with chromunity ids $chid using community detection on a graph. 

Given a GRanges of monomers labeled by concatemer id $cid
}
