% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/chromunity.R
\name{Covariate}
\alias{Covariate}
\title{title}
\value{
Covariate object that can be passed directly to the Chromunity object constructor
}
\description{
Stores Covariates for passing to synergy or annotate function

These are GRanges of interval or numeric type.  The numeric can be optionally log transformed while the interval covariates can be
count (ie count the total number of intervals in the binset) or width based (count the total width in the binset)
}
\author{
Marcin Imielinski
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-new}{\code{Covariate$new()}}
\item \href{#method-merge}{\code{Covariate$merge()}}
\item \href{#method-seqlevels}{\code{Covariate$seqlevels()}}
\item \href{#method-subset}{\code{Covariate$subset()}}
\item \href{#method-toList}{\code{Covariate$toList()}}
\item \href{#method-print}{\code{Covariate$print()}}
\item \href{#method-clone}{\code{Covariate$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-new"></a>}}
\if{latex}{\out{\hypertarget{method-new}{}}}
\subsection{Method \code{new()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Covariate$new(
  data = NULL,
  field = as.character(NA),
  name = as.character(NA),
  log = FALSE,
  count = TRUE,
  pad = 0,
  type = "numeric",
  signature = as.character(NA),
  na.rm = as.logical(NA),
  grep = as.logical(NA)
)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{name}}{character vector Contains names of the covariates to be created, this should not include the names of any Cov objects passed}

\item{\code{log}}{logical flag specifying whether to log the numeric covariate, only applicable to numeric covariate}

\item{\code{count}}{logical flag specifying whether to count the number of intervals}

\item{\code{pad}}{numeric vector Indicates the width to extend each item in the covarite. e.g. if you have a GRanges covariate with two ranges (5:10) and (20:30) with a pad of 5,
These ranges wil become (0:15) and (15:35)}

\item{\code{type}}{character vector Contains the types of each covariate (numeric, interval, sequencing)}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-merge"></a>}}
\if{latex}{\out{\hypertarget{method-merge}{}}}
\subsection{Method \code{merge()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Covariate$merge(...)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-seqlevels"></a>}}
\if{latex}{\out{\hypertarget{method-seqlevels}{}}}
\subsection{Method \code{seqlevels()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Covariate$seqlevels(...)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-subset"></a>}}
\if{latex}{\out{\hypertarget{method-subset}{}}}
\subsection{Method \code{subset()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Covariate$subset(range, ...)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-toList"></a>}}
\if{latex}{\out{\hypertarget{method-toList}{}}}
\subsection{Method \code{toList()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Covariate$toList(...)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-print"></a>}}
\if{latex}{\out{\hypertarget{method-print}{}}}
\subsection{Method \code{print()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Covariate$print(...)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-clone"></a>}}
\if{latex}{\out{\hypertarget{method-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Covariate$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
