% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/chromunity.R
\name{find_multi_modes}
\alias{find_multi_modes}
\title{Determines if a community has multimodal ditribution of contacts along genome}
\usage{
find_multi_modes(
  granges,
  x.field = "start",
  y.field = "score",
  w = 1,
  distance = 1e+05
)
}
\arguments{
\item{granges}{GRanges output from chromunity function for one community}

\item{x.field}{This is the X axis along which smoothing is done}

\item{y.field}{values to be smoothed}

\item{distance}{numeric genomic distance beyond which a peak is not considered local}

\item{which.gr}{the GRanges for the window of interest}
}
\value{
\code{find_multi_modes} returns boolean value if more than one peak is present
}
\description{
Using loess smoothing, determines if a community has multimodal ditribution of contacts along genome
}
\author{
Aditya Deshpande
}
