% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/chromunity.R
\name{fit}
\alias{fit}
\title{Fit the model}
\usage{
fit(
  annotated.binsets,
  nb = TRUE,
  return.model = TRUE,
  verbose = TRUE,
  maxit = 50
)
}
\arguments{
\item{annotated.binsets}{data.table of annotated binsets with at least the fields $bid (binset id) $setid (subset id) $count $min.dist $med.dist $max.dist $width $cardinality $width +/- additional covariates as well, note: every column in this data.table must be a covariate}

\item{nb}{logical flag specifying whether to run negative binomial vs. Poisson model}

\item{return.model}{logical flag whether to return the model or the scored input}

\item{verbose}{logical flag}
}
\value{
fitted model that can be applied to new cases
}
\description{
Fit model to annotated binsets (outputs of annotate)
}
\author{
Aditya Deshpande, Marcin Imielinski
}
