% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/chromunity.R
\name{glm.nb.fh}
\alias{glm.nb.fh}
\title{glm.nb.fh}
\usage{
glm.nb.fh(
  formula,
  data,
  weights,
  subset,
  na.action,
  start = NULL,
  etastart,
  mustart,
  control = glm.control(...),
  method = "glm.fit",
  model = TRUE,
  x = FALSE,
  y = TRUE,
  contrasts = NULL,
  ...,
  theta = NULL,
  init.theta,
  link = log
)
}
\description{
modified glm.nb that takes a pre-defined theta
}
\author{
Marcin Imielinski
}
