% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/chromunity.R
\name{length.Chrom}
\alias{length.Chrom}
\title{length.Chrom}
\usage{
\method{length}{Chrom}(obj, ...)
}
\arguments{
\item{obj}{Cov object that is passed to the length function}
}
\value{
number of covariates contained in the Cov object as defined by length(Cov$data)
}
\description{
Number of binsets in chromunity object
}
\author{
Zoran Z. Gajic
}
