% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/chromunity.R
\name{length.Chromunity}
\alias{length.Chromunity}
\title{length.Chromunity}
\usage{
\method{length}{Chromunity}(obj, ...)
}
\arguments{
\item{obj}{Covariate object that is passed to the length function}
}
\value{
number of covariates contained in the Covariate object as defined by length(Covariate$data)
}
\description{
Number of binsets in chromunity object
}
\author{
Zoran Z. Gajic
}
