% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/chromunity.R
\name{length.Cov}
\alias{length.Cov}
\title{title}
\usage{
\method{length}{Cov}(obj, ...)
}
\arguments{
\item{obj}{Cov object that is passed to the length function}
}
\value{
number of covariates contained in the Cov object as defined by length(Cov$data)
}
\description{
Overrides the length function 'length(Cov)' for use with Cov
}
\author{
Zoran Z. Gajic
}
