% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/chromunity.R
\name{length.Covariate}
\alias{length.Covariate}
\title{title}
\usage{
\method{length}{Covariate}(obj, ...)
}
\arguments{
\item{obj}{Covariate object that is passed to the length function}
}
\value{
number of covariates contained in the Covariate object as defined by length(Covariate$data)
}
\description{
Overrides the length function 'length(Covariate)' for use with Covariate
}
\author{
Zoran Z. Gajic
}
