% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/chromunity.R
\name{muffle}
\alias{muffle}
\title{muffle

Runs code returning NULL is there is any error}
\usage{
muffle(code, ...)
}
\arguments{
\item{code}{R code to eval while suppressing all errors}

\item{...}{additional tryCatch arguments}
}
\value{
output of evaluated R code or NULL if error
}
\description{
muffle

Runs code returning NULL is there is any error
}
\author{
Marcin Imielinski
}
