% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/chromunity.R
\name{names.Chrom}
\alias{names.Chrom}
\title{title}
\usage{
\method{names}{Chrom}(x)
}
\arguments{
\item{obj}{Cov This is the Covariate whose names we are extracting'}
}
\value{
names of covariates
}
\description{
Overrides the names function for Cov object
}
\author{
Zoran Z. Gajic
}
