% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/chromunity.R
\name{names.Covariate}
\alias{names.Covariate}
\title{title}
\usage{
\method{names}{Covariate}(x)
}
\arguments{
\item{obj}{Covariate This is the Covariate whose names we are extracting'}
}
\value{
names of covariates
}
\description{
Overrides the names function for Covariate object
}
\author{
Zoran Z. Gajic
}
