% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/chromunity.R
\name{parquet2gr}
\alias{parquet2gr}
\title{Parquet files to GRanges}
\usage{
parquet2gr(
  path = NULL,
  col_names = NULL,
  save_path = NULL,
  prefix = "NlaIII_this_sample",
  mc.cores = 5,
  verbose = TRUE
)
}
\arguments{
\item{path}{string Path to a directory containing all pore_c.parquet files to be read in.}

\item{col_names}{vector Column names from parquet files to read in.}

\item{save_path}{string Path to save the gRanges to.}

\item{prefix}{string File prefix}

\item{mc.cores}{integer Number of cores in mclapply (default = 5)}

\item{verbose}{boolean "verbose" flag (default = FALSE)}
}
\value{
GRanges format of the parquet files combined
}
\description{
Converts parquet format from Pore-C snakemake pipeline to a single gRanges object
}
\author{
Aditya Deshpande, Marcin Imielinski
}
