% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/chromunity.R
\name{powerset}
\alias{powerset}
\title{Make powerset}
\usage{
powerset(set, min.k = 1, max.k = 5)
}
\arguments{
\item{set}{vector}

\item{min.k}{minimum cardinality subset to include in power set}

\item{max.k}{maximum cardinality subset to include in power set}
}
\description{
Make powerset
}
\author{
Aditya Deshpande, Marcin Imielinski
}
