% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/chromunity.R
\name{rdens}
\alias{rdens}
\title{Empirical dists}
\usage{
rdens(n, values, width = NULL, kernel = "gaussian")
}
\arguments{
\item{n}{integer number of samples}

\item{values}{vector data to be used for building the kernel}

\item{width}{integer kernel width}

\item{kernel}{string kernel type, use Gaussian}
}
\value{
data.able with values drawn from the kernel
}
\description{
getting empirical distributions for a covariate

Function to sample kernel density from a distribution adapted from
https://stats.stackexchange.com/questions/321542/how-can-i-draw-a-value-randomly-from-a-kernel-density-estimate
}
