% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/chromunity.R
\name{shave_concatemers}
\alias{shave_concatemers}
\title{Optimization for reducing concatemer space}
\usage{
shave_concatemers(concatemers, cthresh = 3, bthresh = 2, verbose = TRUE)
}
\arguments{
\item{concatemers}{GRanges of concatemers with $cid and $binid}

\item{cthresh}{integer minimum concatemer order across binset B}

\item{bthresh}{integer minimum bin support across set C of concatemers}

\item{verbose}{logical flag whether to print diff statements cmessages for each step of iteration}
}
\description{
"Shaves" concatemers and bins to a set C and B, respectively such that every bin in B
has at least bthresh concatemer support and every concatemer in C has at least cthresh
bin-wise order.  This is done by iteratively removing concatemers and bins until a fixed point is reached.
}
