% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/chromunity.R
\name{sscore}
\alias{sscore}
\title{Get expectation based on model}
\usage{
sscore(annotated.binsets, model, verbose = TRUE)
}
\arguments{
\item{annotated.binsets}{data.table of annotated binsets with field $count $min.width $med.width $max.width $cardinality $width +/- additional covariates as well}

\item{model}{model output of fit(), the fields of model must match the columns of annotated.binsets}

\item{verbose}{logical flag}
}
\value{
annotated.binsets with additional fields $p and $count.predicted specifying predicted count at each binset
}
\description{
Takes as input annotated binsets and a model and returns each binset scored according to the model
Meaning it annotates the columns with $p, $p.left, p.right, $count.predicted
}
\author{
Aditya Deshpande, Marcin Imielinski
}
