% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/chromunity.R
\name{[.Chrom}
\alias{[.Chrom}
\alias{c.Chrom}
\title{Subset chromunity}
\usage{
\method{c}{Chrom}(...)

\method{[}{Chrom}(obj, range)
}
\arguments{
\item{obj}{Cov This is the Covariate to be subset}

\item{range}{vector This is the range of Covariates to return, like subsetting a vector. e.g. c(1,2,3,4,5)[3:4] == c(3,4)}
}
\value{
A new Cov object that contains only the Covs within the given range

A new Cov object that contains only the Covs within the given range
}
\description{
Overrides the subset operator x[] for use with Chrom

Overrides the subset operator x[] for use with Chrom
}
\author{
Marcin Imielinski

Marcin Imielinski
}
