% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/chromunity.R
\name{[.Cov}
\alias{[.Cov}
\title{title}
\usage{
\method{[}{Cov}(obj, range)
}
\arguments{
\item{obj}{Cov This is the Cov to be subset}

\item{range}{vector This is the range of Covs to return, like subsetting a vector. e.g. c(1,2,3,4,5)[3:4] == c(3,4)}
}
\value{
A new Cov object that contains only the Covs within the given range
}
\description{
Overrides the subset operator x[] for use with Cov to allow for vector like subsetting
}
\author{
Zoran Z. Gajic
}
