% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/chromunity.R
\name{synergy}
\alias{synergy}
\title{Apply synergy}
\usage{
synergy(
  binsets,
  concatemers,
  background.binsets = NULL,
  model = NULL,
  covariates = NULL,
  annotated.binsets = NULL,
  k = 5,
  gg = NULL,
  theta = NULL,
  mc.cores = 5,
  verbose = TRUE,
  resolution = 50000,
  maxit = 50
)
}
\arguments{
\item{binsets}{data.table of bins with fields seqnames, start, end, and $bid specifying binset id}

\item{concatemers}{data.table of monomers with fields seqnames, start, end, and $cid specifying concatemer id, which will be counted across each binset}

\item{background.binsets}{background binsets drawn from output of background or similar, if NULL and model is NULL will be computed [NULL]}

\item{model}{model to fit to binsets / concatemers, if NULL then model is fit after annotating background.binsets, model features must be compatible with provided covariates}

\item{covariates}{covariates to use when annotating binsets and background.binsets}

\item{k}{cardinality k to which to limit k-power set for each binset or background.binset}

\item{gg}{gGraph around which to compute bin to bin distances in binsets and also extract copy numbers using node / edge feature $cn}

\item{mc.cores}{threads to parallelize}

\item{verbose}{logical flag whether to fit}
}
\description{
Applies synergy model by either taking as input (or training) a background model using a set of providved (or generated) background.binsets and testing a set of binmers against a set of concatemers.  The model can be generated using a set of covariates.
}
\author{
Aditya Deshpande, Marcin Imielinski
}
