% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/chromunity.R
\name{chromunity}
\alias{chromunity}
\title{Discovery of communities in Pore-C concatemers}
\usage{
chromunity(
  this.pc.gr,
  k.knn = 10,
  k.min = 1,
  tiles,
  which.gr = which.gr,
  filter_local_number = FALSE,
  filter_local_thresh = NULL,
  take_sub_sample = FALSE,
  frac = 0.25,
  seed.n = 154
)
}
\arguments{
\item{this.pc.gr}{is the GRanges with each row as Pore-C monomer and must have a cmetadata column with corresponding concatemer annotation. The column should be called "read_idx"}

\item{k.knn}{numeric threshoold on k nearest concatemer neighbors to be used to create the graph.}

\item{k.min}{numeric the threshold to number of concatemers pairs to be considered "similar"}

\item{tiles}{GRanges object dividing the genome in a fixed size bin to be defined by user}

\item{which.gr}{the GRanges for the window of interest}

\item{filter_local_number}{boolean (default == FALSE) filters out concatemers in the window that fall below filter_local_thresh length}

\item{filter_local_thresh}{(default == NULL) numeric minimum length of concatemer to be considered for downstream analyses. To be set if filter_local_number == TRUE.}

\item{take_sub_sample}{take a sub sample of concatemers. A random sample of fraction frac is taken}

\item{frac}{fraction of concatemer to be subsampled. To be set if take_sub_sample == TRUE.}

\item{verbose}{boolean (default == TRUE). Outputs progress.}
}
\value{

}
\description{
This function takes in a GRanges with each row as Pore-C monomer with a metadata of the corresponding concatemer id
}
\author{
Aditya Deshpande
}
